requirements.txt
websockets>=11.0.1
python-dotenv==1.0.0